chore: divide brushes type
