+++ apps/chart/src/brushes/spectrumLegend.ts
@@ -1,8 +1,9 @@
 import { padding } from '@src/store/layout';
 import { getTextWidth, getMaxLengthLabelWidth } from '@src/helpers/calculator';
-import { label, bubbleLabel, LabelStyle } from '@src/brushes/label';
+import { label, bubbleLabel } from '@src/brushes/label';
 import { SpectrumLegendModel, SpectrumLegendTooltipModel } from '@t/components/spectrumLegend';
 import { getBubbleArrowPoints } from './dataLabel';
+import { LabelStyle } from '@t/brushes';
 
 export const SPECTRUM_LEGEND_LABEL_HEIGHT = 12;
 export const spectrumLegendBar = {
